refactor(tests): dissolve index.test.ts into dedicated component files
